Rename indexer lucene client packages to be consistent with other clients
